all basic tests in place
