+++ orijenpy/__init__.py
@@ -1,3 +1,4 @@
+"""Package level imports"""
 from .session import Session as session
 from .ns import NS as ns
 from .cred import APIcred as apicred
