+++ tests/test_namespace_role.py
@@ -9,10 +9,3 @@ class TestNSrole:
     def test_list(self, test_session):
         """Method to test list()"""
         assert len(nsrole(test_session).list()['items'])>0
-
-    def test_get(self, test_session):
-        """Method to test get()"""
-        ns_role = nsrole(test_session).list()['items'][0]
-        ns = ns_role['namespace']
-        name = ns_role['name']
-        assert isinstance(nsrole(test_session).get(ns, name), dict)
